+++ pkg/apis/core/v1beta1/helper/helper_managedseed.go
@@ -1,0 +1,166 @@
+// SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package helper
+
+import (
+	"errors"
+	"strconv"
+	"strings"
+
+	"k8s.io/apimachinery/pkg/util/validation/field"
+)
+
+// ManagedSeedAPIServer contains the configuration of a ManagedSeed API server.
+type ManagedSeedAPIServer struct {
+	Replicas   *int32
+	Autoscaler *ManagedSeedAPIServerAutoscaler
+}
+
+// ManagedSeedAPIServerAutoscaler contains the configuration of a ManagedSeed API server autoscaler.
+type ManagedSeedAPIServerAutoscaler struct {
+	MinReplicas *int32
+	MaxReplicas int32
+}
+
+func parseInt32(s string) (int32, error) {
+	i64, err := strconv.ParseInt(s, 10, 32)
+	if err != nil {
+		return 0, err
+	}
+	return int32(i64), nil
+}
+
+func getFlagsAndSettings(annotation string) (map[string]struct{}, map[string]string) {
+	var (
+		flags    = make(map[string]struct{})
+		settings = make(map[string]string)
+	)
+
+	for _, fragment := range strings.Split(annotation, ",") {
+		parts := strings.SplitN(fragment, "=", 2)
+		if len(parts) == 1 {
+			flags[fragment] = struct{}{}
+			continue
+		}
+		settings[parts[0]] = parts[1]
+	}
+
+	return flags, settings
+}
+
+func parseManagedSeedAPIServer(settings map[string]string) (*ManagedSeedAPIServer, error) {
+	apiServerAutoscaler, err := parseManagedSeedAPIServerAutoscaler(settings)
+	if err != nil {
+		return nil, err
+	}
+
+	replicasString, ok := settings["apiServer.replicas"]
+	if !ok && apiServerAutoscaler == nil {
+		return nil, nil
+	}
+
+	var apiServer ManagedSeedAPIServer
+
+	apiServer.Autoscaler = apiServerAutoscaler
+
+	if ok {
+		replicas, err := parseInt32(replicasString)
+		if err != nil {
+			return nil, err
+		}
+
+		apiServer.Replicas = &replicas
+	}
+
+	return &apiServer, nil
+}
+
+func parseManagedSeedAPIServerAutoscaler(settings map[string]string) (*ManagedSeedAPIServerAutoscaler, error) {
+	minReplicasString, ok1 := settings["apiServer.autoscaler.minReplicas"]
+	maxReplicasString, ok2 := settings["apiServer.autoscaler.maxReplicas"]
+	if !ok1 && !ok2 {
+		return nil, nil
+	}
+	if !ok2 {
+		return nil, errors.New("apiSrvMaxReplicas has to be specified for ManagedSeed API server autoscaler")
+	}
+
+	var apiServerAutoscaler ManagedSeedAPIServerAutoscaler
+
+	if ok1 {
+		minReplicas, err := parseInt32(minReplicasString)
+		if err != nil {
+			return nil, err
+		}
+
+		apiServerAutoscaler.MinReplicas = &minReplicas
+	}
+
+	maxReplicas, err := parseInt32(maxReplicasString)
+	if err != nil {
+		return nil, err
+	}
+
+	apiServerAutoscaler.MaxReplicas = maxReplicas
+
+	return &apiServerAutoscaler, nil
+}
+
+func validateManagedSeedAPIServer(apiServer *ManagedSeedAPIServer, fldPath *field.Path) field.ErrorList {
+	allErrs := field.ErrorList{}
+
+	if apiServer.Replicas != nil && *apiServer.Replicas < 1 {
+		allErrs = append(allErrs, field.Invalid(fldPath.Child("replicas"), *apiServer.Replicas, "must be greater than 0"))
+	}
+	if apiServer.Autoscaler != nil {
+		allErrs = append(allErrs, validateManagedSeedAPIServerAutoscaler(apiServer.Autoscaler, fldPath.Child("autoscaler"))...)
+	}
+
+	return allErrs
+}
+
+func validateManagedSeedAPIServerAutoscaler(autoscaler *ManagedSeedAPIServerAutoscaler, fldPath *field.Path) field.ErrorList {
+	allErrs := field.ErrorList{}
+
+	if autoscaler.MinReplicas != nil && *autoscaler.MinReplicas < 1 {
+		allErrs = append(allErrs, field.Invalid(fldPath.Child("minReplicas"), *autoscaler.MinReplicas, "must be greater than 0"))
+	}
+	if autoscaler.MaxReplicas < 1 {
+		allErrs = append(allErrs, field.Invalid(fldPath.Child("maxReplicas"), autoscaler.MaxReplicas, "must be greater than 0"))
+	}
+	if autoscaler.MinReplicas != nil && autoscaler.MaxReplicas < *autoscaler.MinReplicas {
+		allErrs = append(allErrs, field.Invalid(fldPath.Child("maxReplicas"), autoscaler.MaxReplicas, "must be greater than or equal to `minReplicas`"))
+	}
+
+	return allErrs
+}
+
+func setDefaults_ManagedSeedAPIServer(apiServer *ManagedSeedAPIServer) {
+	if apiServer.Replicas == nil {
+		three := int32(3)
+		apiServer.Replicas = &three
+	}
+	if apiServer.Autoscaler == nil {
+		apiServer.Autoscaler = &ManagedSeedAPIServerAutoscaler{
+			MaxReplicas: 3,
+		}
+	}
+
+	setDefaults_ManagedSeedAPIServerAutoscaler(apiServer.Autoscaler)
+}
+
+func minInt32(a int32, b int32) int32 {
+	if a < b {
+		return a
+	}
+	return b
+}
+
+func setDefaults_ManagedSeedAPIServerAutoscaler(autoscaler *ManagedSeedAPIServerAutoscaler) {
+	if autoscaler.MinReplicas == nil {
+		minReplicas := minInt32(3, autoscaler.MaxReplicas)
+		autoscaler.MinReplicas = &minReplicas
+	}
+}
